generalized many functions from models/sample
