+++ cache/models/samples.js
@@ -74,120 +74,6 @@ function cleanAddAspectToSample(sampleObj, aspectObj) {
 }
 
 /**
- * Apply wildcard filter on sample array of keys or objects. For each entry,
- * if given property exists for sample, apply regex to the property value,
- * else if, the property is 'name', then the function was called before getting
- * obj, hence apply regex filter on sample name.
- * @param  {Array}  sampArr - Array of sample keys or sample objects
- * @param  {String}  prop  - Property name
- * @param  {String} propExpr - Wildcard expression
- * @returns {Array} - Filtered array
- */
-function filterByFieldWildCardExpr(sampArr, prop, propExpr) {
-  // regex to match wildcard expr, i option means case insensitive
-  const escapedExp = propExpr.split('_').join('\\_')
-                      .split('|').join('\\|').split('.').join('\\.');
-
-  const re = new RegExp('^' + escapedExp.split('*').join('.*') + '$', 'i');
-  return sampArr.filter((sampEntry) => {
-    if (sampEntry[prop]) { // sample object
-      return re.test(sampEntry[prop]);
-    } else if (prop === sampFields.NAME) { // sample key
-      const sampName = sampleStore.getNameFromKey(sampEntry);
-      return re.test(sampName);
-    }
-
-    return false;
-  });
-}
-
-/**
- * Apply filters on sample keys list
- * @param  {Array} sampKeysArr - Sample key names array
- * @param  {Object} opts - Filter options
- * @returns {Array} - Filtered sample keys array
- */
-function applyFiltersOnSampKeys(sampKeysArr, opts) {
-  let resArr = sampKeysArr;
-
-  // apply limit and offset if no sort order defined
-  if (!opts.order) {
-    resArr = modelUtils.applyLimitAndOffset(opts, sampKeysArr);
-  }
-
-  // apply wildcard expr on name, if specified
-  if (opts.filter && opts.filter.name) {
-    const filteredKeys = filterByFieldWildCardExpr(
-      resArr, sampFields.NAME, opts.filter.name
-    );
-    resArr = filteredKeys;
-  }
-
-  return resArr;
-}
-
-/**
- * Apply field list filter.
- * @param  {Object} sample - Sample object
- * @param  {Array} attributes - Sample fields array
- */
-function applyFieldListFilter(sample, attributes) {
-  // apply field list filter
-  Object.keys(sample).forEach((sampField) => {
-    if (!attributes.includes(sampField)) {
-      delete sample[sampField];
-    }
-  });
-}
-
-/**
- * Apply filters on sample array list
- * @param  {Array} sampObjArray - Sample objects array
- * @param  {Object} opts - Filter options
- * @returns {Array} - Filtered sample objects array
- */
-function applyFiltersOnSampObjs(sampObjArray, opts) {
-  let filteredSamples = sampObjArray;
-
-  // apply wildcard expr if other than name because
-  // name filter was applied before redis call
-  if (opts.filter) {
-    const filterOptions = opts.filter;
-    Object.keys(filterOptions).forEach((field) => {
-      if (field !== sampFields.NAME) {
-        const filteredKeys = filterByFieldWildCardExpr(
-          sampObjArray, field, filterOptions[field]
-        );
-        filteredSamples = filteredKeys;
-      }
-    });
-  }
-
-  // sort and apply limits to samples
-  if (opts.order) {
-    const sortedSamples = modelUtils.sortByOrder(filteredSamples, opts.order);
-    filteredSamples = sortedSamples;
-
-    const slicedSampObjs = modelUtils.applyLimitAndOffset(opts, filteredSamples);
-    filteredSamples = slicedSampObjs;
-  }
-
-  return filteredSamples;
-}
-
-/**
- * Remove extra fields from query body object.
- * @param  {Object} qbObj - Query body object
- */
-function cleanQueryBodyObj(qbObj) {
-  Object.keys(qbObj).forEach((qbField) => {
-    if (!sampleFieldsArr.includes(qbField)) {
-      delete qbObj[qbField];
-    }
-  });
-}
-
-/**
  * Create properties array having and fields and values need to be set to
  * update/create sample. Value is empty string if new object being created.
  * If some value, then calculate status. Update status, previous status and
@@ -198,7 +84,7 @@ function cleanQueryBodyObj(qbObj) {
  * @param  {Object} aspectObj - Aspect object
  */
 function createSampHsetCommand(qbObj, sampObj, aspectObj) {
-  cleanQueryBodyObj(qbObj); // remove extra fields
+  modelUtils.cleanQueryBodyObj(qbObj, sampleFieldsArr); // remove extra fields
   let value;
   if (qbObj[sampFields.VALUE]) {
     value = qbObj[sampFields.VALUE];
@@ -565,7 +451,7 @@ module.exports = {
         });
       }
 
-      cleanQueryBodyObj(reqBody);
+      modelUtils.cleanQueryBodyObj(reqBody, sampleFieldsArr);
       aspectObj = sampleStore.arrayStringsToJson(
         aspObj, constants.fieldsToStringify.aspect
       );
@@ -657,7 +543,7 @@ module.exports = {
         });
       }
 
-      cleanQueryBodyObj(reqBody); // remove extra fields
+      modelUtils.cleanQueryBodyObj(reqBody, sampleFieldsArr); // remove extra fields
       reqBody.name = sampleName; // set name
 
       let value = '';
@@ -754,7 +640,7 @@ module.exports = {
       return modelUtils.checkWritePermission(aspectObj, currSampObj, userName);
     })
     .then(() => {
-      cleanQueryBodyObj(reqBody);
+      modelUtils.cleanQueryBodyObj(reqBody, sampleFieldsArr);
       let value = '';
       if (reqBody.value) {
         value = reqBody.value;
@@ -846,7 +732,7 @@ module.exports = {
 
       // apply fields list filter
       if (opts.attributes) {
-        applyFieldListFilter(sample, opts.attributes);
+        modelUtils.applyFieldListFilter(sample, opts.attributes);
       }
 
       // clean and attach aspect to sample, add api links as well
@@ -884,7 +770,8 @@ module.exports = {
     return redisClient.sortAsync(constants.indexKey.sample, 'alpha')
     .then((allSampKeys) => {
       const commands = [];
-      const filteredSampKeys = applyFiltersOnSampKeys(allSampKeys, opts);
+      const filteredSampKeys = modelUtils
+        .applyFiltersOnSampKeys(allSampKeys, opts);
 
       // add to commands
       filteredSampKeys.forEach((sampKey) => {
@@ -909,12 +796,12 @@ module.exports = {
         sampAspectMap[redisResponses[num].name] = redisResponses[num + ONE];
       }
 
-      const filteredSamples = applyFiltersOnSampObjs(samples, opts);
+      const filteredSamples = modelUtils.applyFiltersOnSampObjs(samples, opts);
       filteredSamples.forEach((sample) => {
 
         const sampName = sample.name;
         if (opts.attributes) { // delete sample fields, hence no return obj
-          applyFieldListFilter(sample, opts.attributes);
+          modelUtils.applyFieldListFilter(sample, opts.attributes);
         }
 
         // attach aspect to sample
